backend/requirements.txt
fastapi
sqlmodel
sqlalchemy
uvicorn
requests
python-dotenv